[nextion] Extract common `upload_end_` function to shared file
Refactors the Nextion TFT upload implementation by extracting the common `upload_end_` function from both Arduino and ESP-IDF specific files into a new shared `nextion_upload_common.cpp` file.

### Changes
- **Added**: `esphome/components/nextion/nextion_upload_common.cpp` - Contains the shared `upload_end_` function
- **Modified**: `nextion_upload_arduino.cpp` - Removed duplicate `upload_end_` implementation
- **Modified**: `nextion_upload_idf.cpp` - Removed duplicate `upload_end_` implementation

### Benefits
- **DRY Principle**: Eliminates code duplication between Arduino and ESP-IDF implementations
- **Maintainability**: Single location for upload completion logic reduces maintenance overhead
- **Consistency**: Ensures identical behavior across all platforms for upload finalization

### Technical Details
The `upload_end_` function handles the final steps of TFT upload including:
- Success/failure logging
- Device reboot on successful upload
- Baud rate restoration on failure
- Cleanup of upload state flags

This function was identical in both platform implementations, making it an ideal candidate for extraction to a common file.
