+++ esphome/components/nextion/nextion_upload_arduino.cpp
@@ -335,31 +335,6 @@ bool Nextion::upload_tft(uint32_t baud_rate, bool exit_reparse) {
   return upload_end_(true);
 }
 
-bool Nextion::upload_end_(bool successful) {
-  ESP_LOGD(TAG, "TFT upload done: %s", YESNO(successful));
-
-  if (successful) {
-    ESP_LOGD(TAG, "Restart");
-    delay(1500);  // NOLINT
-    App.safe_reboot();
-    delay(1500);  // NOLINT
-  } else {
-    ESP_LOGE(TAG, "TFT upload failed");
-
-    this->is_updating_ = false;
-    this->ignore_is_setup_ = false;
-
-    uint32_t baud_rate = this->parent_->get_baud_rate();
-    if (baud_rate != this->original_baud_rate_) {
-      ESP_LOGD(TAG, "Baud back: %" PRIu32 "->%" PRIu32, baud_rate, this->original_baud_rate_);
-      this->parent_->set_baud_rate(this->original_baud_rate_);
-      this->parent_->load_settings();
-    }
-  }
-
-  return successful;
-}
-
 #ifdef USE_ESP8266
 WiFiClient *Nextion::get_wifi_client_() {
   if (this->tft_url_.compare(0, 6, "https:") == 0) {
